+++ envdata/admin.py
@@ -1,8 +1,9 @@
 import site
 
 from django.contrib import admin
-from envdata.models import Emission,FuelEmission
+from envdata.models import Emission, FuelEmission
 
 # Register your models here.
 admin.site.register(Emission)
 admin.site.register(FuelEmission)
+