+++ envdata/forms.py
@@ -1,4 +1,8 @@
-from envdata.models import Emission, FuelEmission
+from envdata.models import (Emission,
+                            FuelEmission,
+                            Sf6Emission,
+                            RefrigerantEmission,
+                            EnergyAcquisition)
 
 from django.forms import ModelForm
 
@@ -8,10 +12,28 @@ class EmissionForm(ModelForm):
         model = Emission
         fields = '__all__'
         labels = {'emission_type': 'Emission Type',
-                  'emission_scope' : 'Emission Scope'}
+                  'emission_scope': 'Emission Scope'}
 
 
 class FuelEmissionForm(ModelForm):
     class Meta:
         model = FuelEmission
+        fields = '__all__'
+
+
+class Sf6EmissionForm(ModelForm):
+    class Meta:
+        model = Sf6Emission
+        fields = '__all__'
+
+
+class RefrigerantEmissionForm(ModelForm):
+    class Meta:
+        model = RefrigerantEmission
+        fields = '__all__'
+
+
+class EnergyAcquisitionForm(ModelForm):
+    class Meta:
+        model = EnergyAcquisition
         fields = '__all__'